Hide header section links on Android
