+++ bin/build-script
@@ -13,9 +13,9 @@ const isProduction = process.env.ELEVENTY_ENV === 'production';
 // some other scripts).
 const inputFiles = [
   'node_modules/focus-visible/dist/focus-visible.min.js',
-  'src/assets/js/setup-dropdown-menu.js',
   'src/assets/js/new-tab-links.js',
   'node_modules/ua-parser-js/dist/ua-parser.min.js',
+  'src/assets/js/header.js', // requires UAParser
   'node_modules/addons-moz-compare/src/index.js',
   'src/assets/js/analytics.js',
   'src/assets/js/dynamic-addon-cards.js',
