+++ src/assets/css/blog-utils-fixes.scss
@@ -1,0 +1,20 @@
+// This file contains changes for the markup/style generated by
+// `addons-frontend-blog-utils`. Ideally, it should be nearly empty ;-)
+
+// stylelint-disable-next-line selector-class-pattern
+.Header--no-section-links {
+  // This is needed to vertically align the Firefox logo.
+  //
+  // stylelint-disable-next-line selector-class-pattern
+  .Header-content {
+    margin-top: 39px;
+
+    @include respond-to(m) {
+      margin-top: 48px;
+    }
+
+    @include respond-to(xl) {
+      margin-top: 0;
+    }
+  }
+}
