+++ src/assets/css/styles.scss
@@ -16,6 +16,7 @@
 @import './header';
 @import './blogposts';
 @import './blog';
+@import './blog-utils-fixes';
 
 * {
   box-sizing: border-box;
